requirements.txt
alpaca-py>=0.21.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
xgboost>=2.0.3
lightgbm>=4.3.0
ta>=0.11.0
pyyaml>=6.0.1
SQLAlchemy>=2.0.30
tqdm>=4.66.4
matplotlib>=3.8.4
plotly>=5.22.0
streamlit>=1.36.0
transformers>=4.43.0
requests>=2.32.5
